chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Sogo Kato", email = "sogo.kato@outlook.com" }
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "line-bot-sdk>=3.11.0",
    "pandas>=2.2.2",
    "openai>=1.30.1",
    "pydantic-settings>=2.2.1",
    "numpy>=1.26.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/chatbot"]
